structure : main페이지 위치 이동
